requirements.txt
viam-sdk==0.40.0

typing-extensions

Pillow